src/main/webapp/app/entities/danh-muc-faq/delete/danh-muc-faq-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucFaqService } from '../service/danh-muc-faq.service';

import { DanhMucFaqDeleteDialogComponent } from './danh-muc-faq-delete-dialog.component';

describe('DanhMucFaq Management Delete Component', () => {
  let comp: DanhMucFaqDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucFaqDeleteDialogComponent>;
  let service: DanhMucFaqService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucFaqDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucFaqDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucFaqDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucFaqService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
